finished adding custom workouts so the intent is recieved and stored created and implimented workoutListAdapter to display the workout's added modified the Exercise file so it is also a parcelable class modified the intent sent to WorkoutExercises and added the list adapter modified the footer for workoutExercises TAGS are mostly all under Mine-FileName Removed the clear() function from the onDataSetChanged
TODO
We still need functions to load all of the data from the database to the adapters. I think that we should only handle this in onStart and onPause so that the database is only accessed to load/update data when the activities are started and finished. It could be nice to change the exercise list for addWorkoutActivity to have TextViews instead of EditTexts. It would be more code and another adapter but would look a lot nicer. Maybe we can find a way to have a generic adapter using a variable for the R.id.listView value as long as the corresponding weight, sets, and reps views have the same names.

I think it might be time to write tests, or add workouts as defaults to the DB. Writing in a new workout every time was not fun.